📝(docs) update readme and a first draft of the documentation
Github pages being activated, this documentation repository is more user
friendly and can be read by all types of users interested in this topic.
The content is iteratively adapted. A first draft (that will be
restructured) is proposed to show for the working group meeting what can
be done with the github pages on this repo.